+++ CHANGELOG.md
@@ -16,14 +16,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [10.1.0] - 2022-08-26
+
 ### Added
 
 - [SIL.Core] Added SIL.PlatformUtilities.Platform.IsFlatpak property.
 - [SIL.Core.Desktop] Added Testing channel to UpdateSettings.
 
 ### Fixed
-
+- [SIL.Archiving] Fixed formatting of DateTimes
 - [SIL.Core] Fixed SIL.IO.PathUtilities.DeleteToRecycleBin and .GetDefaultFileManager to work in a flatpak environment.
+- [SIL.Lexicon] Fixed crash caused by incorrect processing of keyboard data
+- [SIL.Scripture] Fixed SIL.Scripture.MultilingScrBooks.VerseRefRegex to make punctuation more specific
 - [SIL.Windows.Forms] Fixed ImageToolbox.ImageGallery.ImageCollectionManager.FromStandardLocations to work in a flatpak environment.
 - [SIL.WritingSystems] Fixed SLDR initialization for users with European number formats.